# services/proj_wesbank/backend/src/app/api/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Wesbank API",
    "description": "Your Site API ⚡",
    "termsOfService": "https://www.YourSite.co.za/termsandconditions.html",
    "contact": { "url": "https://www.YourSite.co.za/#contact" },
    "version": "1.5.0"
  },
  "paths": {
...
    "/v0/sho002_get_tables_invoice_status_orders_excep": {
      "post": {
        "tags": ["invoice status"],
        "summary": "Sho002 Get Tables Invoice Status Orders Excep",
        "description": "get tables invoice status orders excep",
        "operationId": "sho002_get_tables_invoice_status_orders_excep_v0_sho002_get_tables_invoice_status_orders_excep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/v0/sho002_get_tables_invoice_status_completed_invoices": {
      "post": {
        "tags": ["invoice status"],
        "summary": "Sho002 Get Tables Invoice Status Completed Invoices",
        "description": "get tables invoice status completed invoices",
        "operationId": "sho002_get_tables_invoice_status_completed_invoices_v0_sho002_get_tables_invoice_status_completed_invoices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_tables_invoice_status_accrual_invoices": {
      "post": {
        "tags": ["invoice status"],
        "summary": "Sho002 Get Tables Invoice Status Accrual Invoices",
        "description": "get tables invoice status accrual invoices",
        "operationId": "sho002_get_tables_invoice_status_accrual_invoices_v0_sho002_get_tables_invoice_status_accrual_invoices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/v0/sho002_get_tables_invoice_status_invoice_excep": {
      "post": {
        "tags": ["invoice status"],
        "summary": "Sho002 Get Tables Invoice Status Invoices Excep",
        "description": "get tables invoice status invoice excep",
        "operationId": "sho002_get_tables_invoice_status_invoices_excep_v0_sho002_get_tables_invoice_status_invoice_excep_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/v0/sho002_get_invoice_status_sankey": {
      "post": {
        "tags": ["invoice status"],
        "summary": "Sho002 Get Invoice Status Sankey",
        "description": "Sankey of suppliers invoice status",
        "operationId": "sho002_get_invoice_status_sankey_v0_sho002_get_invoice_status_sankey_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "title": "Formvalues", "type": "object" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "token" } }
      }
    }
  }
}